requirements.txt
pykiwoom
notion-client
pyyaml
beautifulsoup4
requests
feedparser
PyQt5
pandas
